Restructured test package
