+++ src/test/java/io/github/programminglife2016/pl1_2016/parser/metadata/MetaDataParser.java
@@ -1,0 +1,8 @@
+package io.github.programminglife2016.pl1_2016.parser.metadata;
+
+/**
+ * Created by ravishivam on 12-5-16.
+ */
+public class MetaDataParser {
+
+}
