+++ src/test/java/io/github/programminglife2016/pl1_2016/parser/nodes/SegmentCollectionTest.java
@@ -1,6 +1,6 @@
 //CHECKSTYLE.OFF: MagicNumber
 
-package io.github.programminglife2016.pl1_2016.parser;
+package io.github.programminglife2016.pl1_2016.parser.nodes;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonParser;
