+++ src/test/java/io/github/programminglife2016/pl1_2016/parser/nodes/SegmentListTest.java
@@ -1,8 +1,7 @@
 //CHECKSTYLE.OFF: MagicNumber
 
-package io.github.programminglife2016.pl1_2016.parser;
+package io.github.programminglife2016.pl1_2016.parser.nodes;
 
-import io.github.programminglife2016.pl1_2016.parser.nodes.NodeList;
 import org.junit.Before;
 
 /**
